pyproject.toml
[tool.poetry]
name = "fixture-scraper"
version = "0.1.0"
description = ""
authors = ["Matt Hadley <hello@matthadley.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.36.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"